feat: refactor temporal get all blocks info
