+++ src/metadata-block-info/domain/repositories/MetadataBlockInfoRepository.ts
@@ -2,7 +2,7 @@ import { MetadataBlockInfoDisplayFormat, MetadataBlockInfo } from '../models/Met
 
 export interface MetadataBlockInfoRepository {
   getByName: (name: string) => Promise<MetadataBlockInfoDisplayFormat | undefined>
-  getByNameTemporal: (name: string) => Promise<MetadataBlockInfo>
+  getAllTemporal: (names: string[]) => Promise<MetadataBlockInfo[]>
   getDisplayedOnCreateByCollectionId: (
     collectionId: number | string
   ) => Promise<MetadataBlockInfo[]>
