+++ src/sections/create-collection/collection-form/metadata-fields-section/metadata-input-level-fields-block/input-levels-table/InputLevelFieldRow.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, useId } from 'react'
 import { Controller, UseControllerProps, useFormContext, useWatch } from 'react-hook-form'
 import cn from 'classnames'
 import { Form } from '@iqss/dataverse-design-system'
-import { ReducedMetadataFieldInfo } from '../../../../useGetAllMetadataBlocksInfoByName'
+import { ReducedMetadataFieldInfo } from '../../../../useGetAllMetadataBlocksInfo'
 import { INPUT_LEVELS_GROUPER, REQUIRED_BY_DATAVERSE_FIELDS } from '../../../CollectionForm'
 import styles from './InputLevelsTable.module.scss'
 import { CollectionFormHelper } from '../../../CollectionFormHelper'
